refactor(app): move internal/bot/* → app/bot; drop dead agentRunner
Moves workflow, listeners, parser, enrich, skill_provider, and
prompt_context (plus their _test.go counterparts) into app/bot/.

Workflow drops its unused agentRunner field, matching NewWorkflow
parameter, and the worker/agent import — the field was assigned
but never read (grep confirmed zero reads across the package).

cmd/agentdock/app.go moves the agentRunner local into the inmem-only
branch so redis mode no longer builds an unused runner. app/go.mod
gains a require+replace on the root module so app/bot can still
import internal/config (same pattern worker/ uses).
