+++ app/bot/enrich.go
