+++ app/bot/parser_test.go
