+++ app/go.mod
@@ -1,11 +1,15 @@
 module github.com/Ivantseng123/agentdock/app
 
-go 1.25
+go 1.25.0
 
-replace github.com/Ivantseng123/agentdock/shared => ../shared
+replace (
+	github.com/Ivantseng123/agentdock => ../
+	github.com/Ivantseng123/agentdock/shared => ../shared
+)
 
 require (
-	github.com/Ivantseng123/agentdock/shared v0.0.0-00010101000000-000000000000
+	github.com/Ivantseng123/agentdock v0.0.0
+	github.com/Ivantseng123/agentdock/shared v0.0.0
 	github.com/slack-go/slack v0.20.0
 	github.com/xuri/excelize/v2 v2.10.1
 )
@@ -14,6 +18,8 @@ require (
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/google/go-github/v60 v60.0.0 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/gorilla/websocket v1.5.3 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
@@ -31,7 +37,8 @@ require (
 	go.yaml.in/yaml/v2 v2.4.2 // indirect
 	golang.org/x/crypto v0.48.0 // indirect
 	golang.org/x/net v0.50.0 // indirect
-	golang.org/x/sys v0.41.0 // indirect
+	golang.org/x/sys v0.43.0 // indirect
 	golang.org/x/text v0.34.0 // indirect
 	google.golang.org/protobuf v1.36.8 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
